+++ .eslintrc.js
@@ -40,7 +40,7 @@ module.exports = {
     "import/prefer-default-export": ["off"],
     "import/no-extraneous-dependencies": [
       "error",
-      { devDependencies: ["**/__stories__/*", "**/__tests__/*", "/test/*"] }
+      { devDependencies: ["**/__stories__/**", "**/__tests__/*", "/test/*"] }
     ]
   }
 };