+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add test coverage reporting with Codecov. (#2322)
 - Fix use of naive datetime in favor of timezone-aware datetime objects. (#2323)
 - Collect email addresses from github logins
+- Allow users to drop files onto the homepage to create a new notebook with
+  those files attached (#2133)
 
 # 0.13.0 (2019-10-01)